--- code/index.html
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-height" intial-scale="1" />
-    <title>The Perfect Form</title>
-    <link rel="stylesheet" href="vendor/styles/normalize.css">
-    <link rel="stylesheet" href="vendor/style.css">
-    <link rel="stylesheet" href="styles/base.css">
-    <link rel="stylesheet" href="styles/layout.css">
-    <link rel="stylesheet" href="styles/modules.css">
-    <script id="article-template" type="text/x-handlebars-template">
-      <header>
-        <h1><a href="{{projectUrl}}">
-          {{title}}
-        </a></h1>
-        <p class="publishedLine">
-          Published on {{publishedOn}}
-        </p>
-        {{{body}}}
-        <p></br>
-        </p>
-      </header>
-    </script>
-  </head>
-  <body>
-    <section id="main-page" class="main-page-content">
-    <header class="site-header clearfix">
-      <section class="heading-group-h1">
-       <h1>
-        Ben Shields
-       </h1>
-       <h2>
-        253 273 1259
-       </h2>
-      </section>
-      <nav class="main-nav">
-        <div class="icon-office">
-        </div>
-        <ul>
-          <li class="tab" data-content="projects"><a href="file:///Users/BigBen/CodeFellows/301/portfolio/code/index.html">Home</a></li>
-          <li class="tab" data-content="about"><a href="">About</a></li>
-          <li><a href="https://www.linkedin.com/in/william-shields-87740582?trk=nav_responsive_tab_profile_pic" target="_blank"> LinkedIn</a></li>
-          <li><a href="https://github.com/iamrobinhood12345" target="_blank"> GitHub</a></li>
-          <li><a href="https://shieldsbooks.wordpress.com/2016/06/06/25/" target="_blank"> Shields Books</a></li>
-        </ul>
-      </nav>
-    </header>
-    <main>
-      <section id="projects" class="tab-content">
-      </section>
-      <section id="about" class="tab-content">
-        <header>
-          <h1>
-          About
-          </h1>
-        </header>
-        <p>Quotes:</br></br></p>
-        <p>"Three things are to be looked to in a building: that it stands in the right spot; that it be securely founded; that it be successfully executed.""</br>
-        -Johann Wolfgang von Goethe<br/><br/>
-        </p>
-        <p>"I believe in a minimum viable product."</br>
-        -Me<br/><br/>
-        </p>
-      </section>
-      </section>
-      <section class="intro-icons">
-        <div class="icon-play3">
-        </div>
-        <div class="icon-volume-mute">
-        </div>
-        <div class="icon-volume-low">
-        </div>
-        <div class="icon-volume-medium">
-        </div>
-        <div class="icon-volume-high">
-        </div>
-      </section>
-    </main>
-    <footer>
-      &nbsp;&nbsp;&copy; 2016 William Benjamin Shields <|-|-|>
-    </footer>
-    <script src="scripts/jquery.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/handlebars.js/4.0.4/handlebars.js"></script>
-    <script src="scripts/blogProjects.js"></script>
-    <script src="scripts/article.js"></script>
-    <script src="scripts/articleView.js"></script>
-  </section>
-  </body>
-</html>
