+++ package.json
@@ -1,0 +1,15 @@
+{
+  "name": "ben's-page",
+  "version": "1.0.0",
+  "description": "MVC for the win",
+  "main": "server.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "bshields23@gmail.com",
+  "license": "MIT",
+  "dependencies": {
+    "express": "^4.13.3",
+    "express-request-proxy": "^2.0.0"
+  }
+}
