+++ scripts/blogProjects.js
